pyproject.toml
[project]
name = "gtfs-realtime-etl"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-aws-scheduler-alpha>=2.186.0a0",
    "aws-cdk-aws-scheduler-targets-alpha>=2.186.0a0",
    "aws-cdk-lib>=2.218.0",
    "constructs>=10.4.2",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.13.2",
]
